requirements.txt
colorlog==6.7.0
pip>=21.3.1,<23.2
pylint==2.8.2
ruff==0.7.2
pre-commit

# Install HA and test dependencies (pytest, coverage)
# To pin the dev container to a specific HA version, set this dependency
# to the adequate version (add `==<version>`) and rebuild the dev container.
# See https://github.com/MatthewFlamm/pytest-homeassistant-custom-component/releases for version mappings.
pytest-homeassistant-custom-component
